Implemented authentication reroute

+++ start-project.py
@@ -115,12 +115,12 @@ def Add_Restaurant():
     elif gAuth.Is_Authenticated():
         return render_template('add-restaurant.html')
     else:
-        return url_for('Show_All_Restaurants')
+        return redirect(url_for('Show_All_Restaurants'))
 
 
 @app.route('/restaurant/<int:rest_id>/edit', methods=['GET', 'POST'])
 def Edit_Restaurant(rest_id):
-    if request.method == 'POST':
+    if request.method == 'POST' and gAuth.Is_Authenticated():
         updated_restaurant = session.query(Restaurant).filter_by(id=rest_id).one()
         updated_restaurant.name = request.form['rest_name']
 
@@ -130,20 +130,23 @@ def Edit_Restaurant(rest_id):
         flash('Successfully updated %s.' % (updated_restaurant.name))
 
         return redirect(url_for('Edit_Restaurant', rest_id=rest_id))
-    
-    rest_data = Get_Restaurant_Data(rest_id)
+    elif gAuth.Is_Authenticated():
+        rest_data = Get_Restaurant_Data(rest_id)
 
-    return render_template('edit-restaurant.html', 
-                            restaurant=rest_data['restaurant'], 
-                            appetizer_items=rest_data['appetizers'], 
-                            drink_items=rest_data['drinks'],
-                            entree_items=rest_data['entrees'],
-                            dessert_items=rest_data['desserts'])
+        return render_template('edit-restaurant.html', 
+                                restaurant=rest_data['restaurant'], 
+                                appetizer_items=rest_data['appetizers'], 
+                                drink_items=rest_data['drinks'],
+                                entree_items=rest_data['entrees'],
+                                dessert_items=rest_data['desserts'])
+
+    else:
+        return redirect(url_for('Show_All_Restaurants'))
 
 
 @app.route('/restaurant/<int:rest_id>/delete', methods=['GET', 'POST'])
 def Delete_Restaurant(rest_id):
-    if request.method == 'POST':
+    if request.method == 'POST' and gAuth.Is_Authenticated():
         restaurant = session.query(Restaurant).filter_by(id=rest_id).one()
         if restaurant != []:
             session.query(MenuItem).filter_by(restaurant_id=rest_id).delete()
@@ -154,16 +157,17 @@ def Delete_Restaurant(rest_id):
         session.commit()
 
         return redirect(url_for('Show_All_Restaurants'))
-    
-    rest_data = Get_Restaurant_Data(rest_id)
+    elif gAuth.Is_Authenticated():
+        rest_data = Get_Restaurant_Data(rest_id)
 
-    return render_template('delete-restaurant.html', 
-                            restaurant=rest_data['restaurant'], 
-                            appetizer_items=rest_data['appetizers'], 
-                            drink_items=rest_data['drinks'],
-                            entree_items=rest_data['entrees'],
-                            dessert_items=rest_data['desserts'])
-
+        return render_template('delete-restaurant.html', 
+                                restaurant=rest_data['restaurant'], 
+                                appetizer_items=rest_data['appetizers'], 
+                                drink_items=rest_data['drinks'],
+                                entree_items=rest_data['entrees'],
+                                dessert_items=rest_data['desserts'])
+    else:
+        return redirect(url_for('Show_All_Restaurants'))
 
 @app.route('/restaurants/all/items/')
 def Show_All_Items():
@@ -174,7 +178,7 @@ def Show_All_Items():
 
 @app.route('/restaurant/<int:rest_id>/add', methods=['GET', 'POST'])
 def Add_Menu_Item(rest_id):
-    if request.method == 'POST':
+    if request.method == 'POST' and gAuth.Is_Authenticated():
         form = request.form
 
         new_menu_item = MenuItem(
@@ -191,14 +195,16 @@ def Add_Menu_Item(rest_id):
         flash("Successfully added %s." % (new_menu_item.name))
 
         return redirect(url_for('Edit_Restaurant', rest_id=rest_id))
+    elif gAuth.Is_Authenticated():
+        restaurant = session.query(Restaurant).filter_by(id=rest_id).one()
 
-    restaurant = session.query(Restaurant).filter_by(id=rest_id).one()
-
-    return render_template('add-menu-item.html', restaurant=restaurant)
+        return render_template('add-menu-item.html', restaurant=restaurant)
+    else:
+        return redirect(url_for('Show_All_Restaurants'))
 
 @app.route('/restaurant/<int:rest_id>/edit/<int:item_id>/', methods=['GET', 'POST'])
 def Edit_Menu_Item(rest_id, item_id):
-    if request.method == 'POST':
+    if request.method == 'POST' and gAuth.Is_Authenticated():
         form = request.form 
 
         updated_item = session.query(MenuItem).filter_by(id=item_id).one()
@@ -214,16 +220,18 @@ def Edit_Menu_Item(rest_id, item_id):
         flash('Successfully updated %s.' % (updated_item.name)) 
 
         return redirect(url_for('Edit_Restaurant', rest_id=rest_id))
+    elif gAuth.Is_Authenticated():
+        restaurant = session.query(Restaurant).filter_by(id=rest_id).one()
+        item = session.query(MenuItem).filter_by(id=item_id).one()
 
-    restaurant = session.query(Restaurant).filter_by(id=rest_id).one()
-    item = session.query(MenuItem).filter_by(id=item_id).one()
-
-    return render_template('edit-menu-item.html', restaurant=restaurant, item=item)
+        return render_template('edit-menu-item.html', restaurant=restaurant, item=item)
+    else:
+        return redirect(url_for('Show_All_Restaurants'))
 
 
 @app.route('/restaurant/<int:rest_id>/delete/<int:item_id>/', methods=['GET', 'POST'])
 def Delete_Menu_Item(rest_id, item_id):
-    if request.method == 'POST':
+    if request.method == 'POST' and gAuth.Is_Authenticated():
         item_to_delete = session.query(MenuItem).filter_by(id=item_id).one()
         if item_to_delete != []:
             session.delete(item_to_delete)
@@ -232,11 +240,13 @@ def Delete_Menu_Item(rest_id, item_id):
             flash('Successfully removed %s.' % (item_to_delete.name))
 
         return redirect(url_for('Edit_Restaurant', rest_id=rest_id))
-    
-    restaurant = session.query(Restaurant).filter_by(id=rest_id).one()
-    item = session.query(MenuItem).filter_by(id=item_id).one()
+    elif gAuth.Is_Authenticated():
+        restaurant = session.query(Restaurant).filter_by(id=rest_id).one()
+        item = session.query(MenuItem).filter_by(id=item_id).one()
 
-    return render_template('delete-menu-item.html', restaurant=restaurant, item=item)
+        return render_template('delete-menu-item.html', restaurant=restaurant, item=item)
+    else:
+        return redirect(url_for('Show_All_Restaurants'))
 
 
 @app.route('/restaurants/JSON')
